pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ebay-research-edge"
version = "0.1.0"
description = "eBay Sold and domestic price comparison tool for Japanese sellers"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "nario", email = "nario@example.com" }
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
